Cyp 1594 2.01.0 minor adjustments (#102)
* fix: :adhesive_bandage: duplicate padding removal in portfoliotokenItem

Removed the extra padding that was in the portfolio token item.

* fix: :adhesive_bandage: options page logo not safe fix

The Cypher logo in the Options page was not safe in all the screens with notches for e.g. Fixed this by adding margin.

* fix: :adhesive_bandage: centered the hide/show icon in Portfolio

The HIDE/SHOW icon and text in the Portfolio Banner is now centered rather than being aligned at the end.