fix: :ambulance: card Imagebackground glitching (#103)
* fix: :ambulance: card Imagebackground glitching

The ImageBackground used for the cards in the Cypher Card page is now replaced an absolute positioned FastImage, which stops the gltiching.

* fix: :art: centered single card in CardCarousel for android

Card Carousel with single card was not centered in android since contentInset doesn't work in android. So paddingHorizontal was added to compensate for that.

* feat: :loud_sound: add log for card transaction info click

An event called card_transaction_info_clicked will be logged when a user clicks on any of the Card transaction to look at the transaction details.