Cyp 1600 card txns loading everytime (#107)
* style: :lipstick: removed unnecessary padding for android

Removed the unecessary padding that was added to compensate for cards not being centered in Android. Removing this because, now the cards are centered.

* refactor: :zap: switched scrollview to flatlist

Changed the ScrollView to Flatlist for the Card Transactions page.

* refactor: :recycle: included the empty component in the flatlist

Instead of checking for the length and conditional rendered, used the flatlist EmptyComponent prop to render empty list.

* fix: :adhesive_bandage: excessive loading on back and forth from card txns and txndetails

Included the TxnDetails page in the list to have a relative tab bar so the re-rendering of the sheet doesn't happen due to change of height.

* feat: :sparkles: add pull to refresh + few changes.

Added pull to refresh to the Card Transactions. Removed Apto Card page navigator. Changed title to headerTitle to have safe title in Cypher Card.