fix: :lipstick: white bg for the txn list when list not full (#123)
* fix: :lipstick: white bg for the txn list when list not full

When the txn list is either empty or not full, the list showed the bgImage. Now added a View that is attached to list items and positioned under to have a white bg. And added white bg for ListEmptyComponent

* fix: :lipstick: small change in positioning

Change in positioning to have white bg under the first element as well.

* fix: :iphone: remove refreshcontrol bg for android, change date filter label

Removed the bg for refreshcontrol in android ( unnecessary and wrong ). Change the date range filter label to date.