CYP-1645-Physical-card-upgrade-screen (#142)
* feat: :sparkles: add new screen for upgrading to physical card

Added a new Upgrade to Physical Card screen that takes in user shipping details. This flow is identical to the dApp.

* fix: :bug: remove console logs

Removed the console logs and added Sentry captures whereever needed.

* refactor: :building_construction: change route to UpgradeScreen

Added a extra card when there's physical card upgrade available. Similar to extra card when activate physical card.

* feat: :sparkles: PR changes

Added a separate screen for OTP like the other parts of the app.