Cyp 1671 non critical card and txndetails bugs (#154)
* fix: :building_construction: reset the navigation stack after cardDetails load

Reset the navigation stack after the cardDetails load to the V2 card screen and the correct params.

This fixes the unexpected navigation where the users were be able to come back to the cardwaitlist screen from the V2 Card screen.

* feat: :adhesive_bandage: remove isSettled from Credit txns, formatted amount for Noble USDC

Formatted the loaded amount with the formatAmount util. Removed the isSettled information from the Credit transactions.

This will break the old version of the apps, when the api response doesn't send isSettled for Credit transactions anymore.