CYP-1685-fund-card-streamlining-and-enhancements (#160)
* fix: :adhesive_bandage: decreased font size for longer amount in card load input

Decreasing font size based on the length of the card load amount.

When clicking on max in card load, and having a token in decimals caused the input to focus at the end. It still does this but now the font size is reduced to show the entire amount.

* feat: :sparkles: display zero fee funding indicator, 1:1 pegging for inputs

Made an indicator to show that a token is Zero fee for card funding. The conversion from Token Value to USD value is 1:1 for the zeroFee tokens.