fix: :ambulance: ethers NUMERIC_FAULT error during fundCard
limited decimal places to contract decimals for the tokenRequired received from the card funding quote. This was previously fixed for gas estimation. Now this has also been fixed for the actual transaction.

This fixes the high decimal precision causing the precision error when even after the gas estimation is succesful for the same tokens required.